fix(yarn): update install command for ci (#305)
Update the CI configuration to use modern yarn.

I've switched us to using the node orb instead of just the cimg/node Docker image. We end up using the same Docker image, just a different route to get there. The orb comes with a node/install-packages helper that wraps the install process and handles caching for us. Therefore, I removed the caching code.

I also lowered the resource_class of the run to medium, since we barely do anything in it. 4GB RAM should be plenty.